submodule_store: add get_submodule_path()
This function will tell the `jj git clone` machinery where to clone to.
It's quite a leaky abstraction (I think it's more appropriate for the
submodule store to hide the details of how cloning is done), but a
better abstraction will require much more refactoring, so let's punt on
that for now in the name of quick experimentation.